chore: 서버 내 보관 돈 단위 '원'으로 통일

+++ payment-service/src/main/java/com/lgcns/bebee/payment/application/usecase/ConfirmPaymentUseCase.java
@@ -78,7 +78,7 @@ public class ConfirmPaymentUseCase implements UseCase<ConfirmPaymentUseCase.Para
         HoneyHistory history = HoneyHistory.create(
                 wallet,
                 temp.getMemberId(),
-                response.totalAmount(),
+                response.totalAmount().longValue(),
                 HoneyHistoryType.CHARGE
         );
         honeyHistoryRepository.save(history);
